pnger.cpp translation unit
