+++ src/image.cpp
@@ -1,6 +1,7 @@
 #include "src/functions.h"
 #include "src/geotiff.h"
 #include "src/mapalloc.h"
+#include "src/pnger.h"
 
 int hist_red[256];
 int hist_grn[256];
