+++ src/mapalloc.h
@@ -3,6 +3,7 @@
 #include <vector>
 
 #ifdef _WIN32
+#define NOMINMAX
 #include <Windows.h>
 #endif
