+++ src/pyramid.cpp
@@ -5,6 +5,8 @@
 #include "sse_mathfun.h"
 #endif
 
+#include "src/pnger.h"
+
 /***********************************************************************
  * Constructor/destructor
  ***********************************************************************/
