+++ src/threadpool.h
@@ -6,6 +6,7 @@
 #include <mutex>
 
 #ifdef _WIN32
+#define NOMINMAX
 #include <Windows.h>
 #endif
